+++ Firestore/core/src/local/lru_garbage_collector.h
@@ -147,6 +147,13 @@ class LruGarbageCollector {
 
   local::LruResults Collect(const LiveQueryMap& live_targets);
 
+  /**
+   * Visible for testing only!
+   */
+  void set_lru_params(LruParams params) {
+    params_ = params;
+  }
+
  private:
   LruResults RunGarbageCollection(const LiveQueryMap& live_targets);